backend/requirements.txt
# Dependências da API (alinhadas ao .venv local em 2026-04-01)
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.2
python-jose[cryptography]==3.3.0
cryptography==43.0.3
loguru==0.7.2
supabase==2.7.4
